Quick-fix modern YouTube timedtext subtitle position
These come from YT with a "position:0%" formatting instruction,
which libass within Kodi interprets as "at the absolute left edge of
the displayable area". For comparison, YouTube's own web player positions
those (arguably wrongly) indented by a fair amount.

This quick-fix unconditionally readjusts this absolute positioning
instruction to "5%" (from the left edge), which looks pleasant to me
and visually matches the border below the subtitles, when set to
"at bottom of video".

This quick-fix is only effective with the "download subtitles before
starting playback" configuration setting turned on. (This is because
with that setting turned off, only the source URL is passed on to Kodi
and it reads subtitle data directly from YouTube content delivery servers.)
In that sense, said configuration options also provides a poor man's way
to enable or disable this hacky fix.

I'm aware that this probably does not meet quality of user experience
expectations... it's what I can do today, it Works For Me(tm) and in
this way it's easier for me to cherry-pick into future releases, as well
as maybe available to everyone else looking for it hard enough ;-)